Bug Fix
Bug: Items could be bought from players inventories while in the shops.
If the player had a diamond sword in their inventory they could buy it
by clicking on it instead of clicking on the one in the shop. Players
couldn't move items around in their inventories while a shop was open.

Fix: Items can only be bought from the shop and players can move items
around in their inventories.